Move ClickHouseDefaults to the clickhouse module
